move login plugin
